Yet more epicycles.  Remove reference to a debug module i'm using.
This would be a good time to note that the last commits added new functionality to the modules:  There are methods that are run befgore the first entry is processed and after the last entry is processed.  This allows for cerrtain operations on histograms to be performed once (e.g. initialising specific histograms dependent on the bank names present, normalising operations etc...)

I have also modified the abort behaviour if a module returns non-zero in processing.  Now all modules will be given a chance to operate on the entry before breaking out.