Use references to std::vector<TXxxPulses*> rather than copying them. As spotted by John, this stops the originals from being Cleared().  The problem occured becasue I copypasted the code to TGlobalData.  Originally this hardly mattered because the vectors got deleted behind the scenes by ROOT, but the change meant the elemets were deleted twice just before that expiry.  This should fix issue #72.
In the long run we should make TGlobalData soley responsible for cleaning itself up, since it must do this any time it get deleted (Obect Orientaion 101!).  Retrospectivly, I think we should have done this in the first place - it isn't actually a risk for alcapana if alcapana never calls the method.