Preliminary changes to data_quality modules
I've updated most of the data quality modules.  the FADC and muSc modules are to be ignored.  Thresholds still need to be updated, and the others still need to be tested.  Updated MODULES file for R15a, include error checks on CAEN, Sync pulses per event in each WFD and TDC, some WFD to TDC correlations and all relevant data_quality plots.