Editted timing routines to improve efficiency
Inverted equation for MTDCWFDMatch, included WFD timing in the TPulseIsland (to be defined by user).  Added WFD time set in MTDCWFDMatch.  Included a check in MTCorrTest and MTCorrWFD to reduce the number of missed correlations.  Moved the tree generation in MTCorrTree to _bor in order to save more than the last event.